chore(0.4): retire features/mcp-server-install (0.3.x binary installer)
Not invoked in 0.4.x (in-process server, no binary). Relocated the two
pieces still imported live: SetupResult -> features/core/types.ts;
BINARY_NAME/INSTALL_PATH/Platform -> features/migration/constants.ts (the
migration UX still detects a leftover 0.3.x binary). Deletes the rest of
the module + its tests, eliminating the automated-review unsafe-any /
unused-var / err-stringify / await-non-Promise warnings sourced from
mcp-server-install/services/*.
