+++ packages/obsidian-plugin/src/features/core/types.ts
@@ -1,0 +1,9 @@
+/**
+ * Result contract for a feature `setup()`. Relocated here from the
+ * retired 0.3.x binary-installer module — `core` is the only live
+ * consumer in the 0.4.x line.
+ */
+export interface SetupResult {
+  success: boolean;
+  error?: string;
+}
